requirements.txt
streamlit==1.24.0
langchain==0.0.200
openai==0.27.8
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
stability-sdk==0.8.0
ollama==0.1.0
python-multipart==0.0.6